backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
sqlalchemy>=2.0.0
pydantic
pydantic-settings
python-dotenv
PyJWT
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
better-exceptions
alembic
python-dateutil
cohere
openai
requests
slowapi